+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## 1.1.2
 - added more Dutch translation (thank you Fabijenna)
 - fixed the *leave distraction free mode* button icon
+- the notes viewed in the distraction free mode now have a viewport width of 60 
+  characters (for [Issue #86](https://github.com/pbek/QOwnNotes/issues/86))
+    - the viewport will resize automatically if the width of the window changes
 
 ## 1.1.1
 - added **Dutch translation** (a big thank you to Fabijenna!)